Improve PHP7 compatibility by allowing symfony/console:^4
The current set of requirements for Peridot makes it difficult
to install on an up-to-date PHP 7 environment. Composer itself
will install with version 4 of Symfony Console, which is not
installable with the config platform of php:5.4.

I have updated to allow symfony/console:^4, but this is still
not ideal as it requires installation using the
--ignore-platform-reqs flag.